requirements.txt
# Core dependencies
numpy>=1.20.0
openai>=1.0.0
tenacity>=8.0.0
pyyaml>=6.0
ase>=3.22.0  # Atomic Simulation Environment
torch
scipy
mace
mace-torch

# Utilities
requests>=2.25.0
tqdm>=4.60.0
pathlib>=1.0.1
typing-extensions>=4.0.0
periodictable
tqdm

# File processing
h5py>=3.6.0  # For HDF5 dataset creation

# Crystallography tools
pymatgen>=2022.0.0  # For advanced crystallography functions

# Visualization 
matplotlib>=3.5.0
seaborn>=0.11.0



